What I worked on
I tried to work on line clearing, and it works for the most part. The big issue I got stuck on was athat if an object moves down or left then the whole block disappears instead of juts the blocks in the full row/column and if the object moves up or right then none of the block sprites disappear, but they stop appearing on the grid.